requirements/base.txt
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-xml==1.4.0
Pillow==6.0.0
psycopg2==2.8.1
